dcerpc: NetBSD compile fixes
* Include standard headers where necessary
* #undef system macros for pthread_xxx routines

CR: dalmeida@likewise.com

(dcerpc: r39717)
(cherry picked from commit 18e2678d643baf2e06d8ca461662cfc1c4bf5fa4)

+++ dcerpc/ncklib/twr_np.c
@@ -197,7 +197,7 @@ PUBLIC void twr_np_lower_flrs_from_sa
     gethostname(&hostname[2], MAXHOSTNAMELEN - 3);
     hostname[MAXHOSTNAMELEN - 1] = '\0';
     for (p = &hostname[2]; *p != '\0'; p++)
-        *p = toupper(*p);
+        *p = toupper((int)*p);
     related_data_size[1] = strlen(hostname) + 1;
     related_data_ptr[1] = (unsigned char*) hostname;
